restructured files and folders again
