--- outreach-mapping-solutions-fe-react/src/components/clientInformation.jsx
@@ -1,3 +1,0 @@
-export default function ClientInformation() {
-  return <h1>Client Information</h1>;
-}
