+++ outreach-mapping-solutions-fe-react/src/components/clientProfileComponents/clientInformation.jsx
@@ -1,0 +1,5 @@
+import ClientBaseRecord from "./clientInformationComponents/clientBaseRecord";
+
+export default function ClientInformation({ clientData }) {
+  return <ClientBaseRecord clientData={clientData} />;
+}
