+++ outreach-mapping-solutions-fe-react/src/components/clientProfileComponents/clientInformationComponents/clientBaseRecord.jsx
@@ -1,0 +1,7 @@
+export default function ClientBaseRecord({ clientData }) {
+  return (
+    <>
+      <p>Name:</p> <span>{clientData.firstName}</span>
+    </>
+  );
+}
